implement video capture:export to mp4

+++ JobSnitch/DataModel/PostingRecord.h
@@ -12,7 +12,7 @@
 @property (nonatomic, strong) NSString * title;
 @property (nonatomic, strong) NSString * descrption;
 @property (nonatomic, strong) NSString * JobCategoryName;             // ?
-@property (nonatomic) int   JobCategoryId;
+@property (nonatomic) NSInteger   JobCategoryId;
 @property (nonatomic, strong) NSString * industry;                   // ?
 @property (nonatomic) BOOL   morningShift;
 @property (nonatomic) BOOL   afternoonShift;
@@ -26,8 +26,8 @@
 @property (nonatomic) BOOL   wantsVideo;
 
 @property (nonatomic) BOOL   isActive;
-@property (nonatomic) int   CompanyId;
-@property (nonatomic) int   JobPostingId;
+@property (nonatomic) NSInteger   CompanyId;
+@property (nonatomic) NSInteger   JobPostingId;
 @property (nonatomic, strong) NSString * JobLocation;
 
 @property (nonatomic, weak) id ownerBusiness;
